Separated server and client Navbar components (facilitates resource sharing and rehydration)
