+++ utils/auth-helpers/server.ts
@@ -15,6 +15,26 @@ export async function redirectToPath(path: string) {
   return redirect(path);
 };
 
+export async function redirectToSignIn() {  
+  return redirect('/signin');
+};
+
+export async function SignOut (formData: FormData) {
+  const pathName = String(formData.get('pathName'));
+  
+  const cookieStore = cookies();
+  const supabase = createClient(cookieStore);
+  const { error } = await supabase.auth.signOut();
+
+  if (error) {
+    return redirect(
+      getErrorRedirect(pathName, 'Hmm... Something went wrong.', 'You could not be signed out.')
+    );
+  }
+
+  return redirect('/signin');
+};
+
 export async function signInWithEmail (formData: FormData) {
   const callbackURL = getURL('/auth/callback');
